backend_api/requirements.txt
asgiref
autopep8
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pycodestyle
PyJWT
pytz
sqlparse
toml
tzdata
django-import-export
django-jazzmin
django-crispy-forms
django-mathfilters
django-ckeditor
django-ckeditor-5
django-taggit
shortuuid
# mysql
# mysqlclient
# mysql-connector
psycopg2-binary



